requirements.txt
python-dotenv~=0.15.0
discord.py~=1.6.0
tensorflow==2.4.1
tensorflow-datasets==4.2.0
pycryptodome==3.10.1
nltk==3.5
numpy==1.19.5